# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KKN Integration Dashboard API",
        "description": "Dokumentasi API dashboard KKN. API ini dipakai untuk integrasi dengan aplikasi project-akhir (portal mahasiswa & dosen) yang mengirim data pendaftaran, dokumen, dan nilai ke dashboard.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Server lokal dashboard"
        }
    ],
    "paths": {
        "/api/kkn/store-from-external": {
            "post": {
                "tags": [
                    "KKN"
                ],
                "summary": "Menerima pendaftaran KKN dari project-akhir",
                "description": "Menyimpan data pendaftaran KKN beserta anggota kelompok dan file proposal (PDF). Dikirim oleh aplikasi project-akhir saat mahasiswa mendaftar.",
                "operationId": "e65e5b1923a2e19f3e4e3c1c8b84cbfe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul_kegiatan",
                                    "mitra",
                                    "lokasi_mitra",
                                    "nama[]",
                                    "nim[]",
                                    "prodi[]",
                                    "peran[]",
                                    "file",
                                    "user_nim"
                                ],
                                "properties": {
                                    "judul_kegiatan": {
                                        "type": "string",
                                        "example": "KKN Tematik Desa Sukamaju",
                                        "maxLength": 255
                                    },
                                    "mitra": {
                                        "type": "string",
                                        "example": "Pemerintah Desa Sukamaju",
                                        "maxLength": 100
                                    },
                                    "lokasi_mitra": {
                                        "type": "string",
                                        "example": "Desa Sukamaju, Kab. Bandung",
                                        "maxLength": 255
                                    },
                                    "nama[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Budi Santoso",
                                            "Siti Aminah"
                                        ]
                                    },
                                    "nim[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "2010001",
                                            "2010002"
                                        ]
                                    },
                                    "prodi[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Teknik Informatika",
                                            "Sistem Informasi"
                                        ]
                                    },
                                    "peran[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "Ketua",
                                                "Anggota"
                                            ]
                                        },
                                        "example": [
                                            "Ketua",
                                            "Anggota"
                                        ]
                                    },
                                    "file": {
                                        "description": "File proposal PDF (maks 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_nim": {
                                        "type": "string",
                                        "example": "2010001",
                                        "maxLength": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pendaftaran berhasil disimpan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Terjadi kesalahan server"
                    }
                }
            }
        },
        "/api/kkn/pendaftar": {
            "get": {
                "tags": [
                    "KKN"
                ],
                "summary": "Daftar semua pendaftar KKN",
                "description": "Mengambil seluruh data pendaftar KKN beserta ringkasan anggota dan status verifikasi.",
                "operationId": "09b4d449029ffff1ad46016db7191d7a",
                "responses": {
                    "200": {
                        "description": "Daftar pendaftar berhasil diambil"
                    },
                    "500": {
                        "description": "Terjadi kesalahan server"
                    }
                }
            }
        },
        "/api/kkn/pendaftar/user/{userNim}": {
            "get": {
                "tags": [
                    "KKN"
                ],
                "summary": "Daftar pendaftar berdasarkan NIM user",
                "description": "Mengambil data pendaftaran KKN milik user tertentu berdasarkan NIM.",
                "operationId": "8fb25a569c5f1cf00edacbe631e0245d",
                "parameters": [
                    {
                        "name": "userNim",
                        "in": "path",
                        "description": "NIM user pendaftar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2010001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pendaftar berhasil diambil"
                    },
                    "500": {
                        "description": "Terjadi kesalahan server"
                    }
                }
            }
        },
        "/api/kkn/pendaftar/{id}": {
            "get": {
                "tags": [
                    "KKN"
                ],
                "summary": "Detail pendaftar KKN",
                "description": "Mengambil detail satu pendaftar KKN beserta anggotanya berdasarkan ID.",
                "operationId": "23faaa47f6fcc871b7a1bae043f17e39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pendaftar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pendaftar berhasil diambil"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan API key mitra di sini untuk endpoint tarik data nilai.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}